+++ src/sx/scripts/setupTests.js
@@ -1,3 +1,4 @@
 // @flow strict
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import '@testing-library/jest-dom/extend-expect';